requirements-test.txt
pytest>=7.4
pyspark>=3.3
delta-spark>=2.4
PyYAML>=6.0
pytest-cov>=4.0
